feat: add the anonymous file upload in form gateway

+++ apps/form-app/src/app/components/DraftForm.tsx
@@ -25,6 +25,7 @@ import {
   formActions,
   metaDataSelector,
   uploadFile,
+  uploadAnonymousFile,
 } from '../state';
 import { userSelector, configSelector } from '../state';
 
@@ -39,6 +40,7 @@ interface DraftFormProps {
   showSubmit: boolean;
   saving: boolean;
   submitting: boolean;
+  anonymousApply?: boolean;
   onChange: ({ data, errors }: { data: unknown; errors?: ValidationError[] }) => void;
   onSubmit: (form: Form) => void;
 }
@@ -94,6 +96,7 @@ export const DraftForm: FunctionComponent<DraftFormProps> = ({
   showSubmit,
   saving,
   submitting,
+  anonymousApply,
   onChange,
   onSubmit,
 }) => {
@@ -121,9 +124,15 @@ export const DraftForm: FunctionComponent<DraftFormProps> = ({
       dispatch(deleteFile({ urn, propertyId }));
     }
 
-    const fileMetaData = (
-      await dispatch(uploadFile({ typeId: FORM_SUPPORTING_DOCS, recordId: form.urn, file, propertyId })).unwrap()
-    ).metadata;
+    const fileMetaData =
+      anonymousApply === true
+        ? (
+            await dispatch(
+              uploadAnonymousFile({ typeId: FORM_SUPPORTING_DOCS, recordId: form?.urn, file, propertyId })
+            ).unwrap()
+          ).metadata
+        : (await dispatch(uploadFile({ typeId: FORM_SUPPORTING_DOCS, recordId: form.urn, file, propertyId })).unwrap())
+            .metadata;
 
     clonedFiles[propertyId] = fileMetaData.urn;
     dispatch(formActions.updateFormFiles(clonedFiles));
@@ -142,7 +151,10 @@ export const DraftForm: FunctionComponent<DraftFormProps> = ({
     const clonedFiles = { ...files };
     const propertyId = getKeyByValue(clonedFiles, file.urn);
 
-    await dispatch(deleteFile({ urn: file.urn, propertyId }));
+    if (anonymousApply !== true) {
+      await dispatch(deleteFile({ urn: file.urn, propertyId }));
+    }
+
     delete clonedFiles[propertyId];
 
     dispatch(formActions.updateFormFiles(clonedFiles));
@@ -165,7 +177,6 @@ export const DraftForm: FunctionComponent<DraftFormProps> = ({
             downloadFile: downloadFormFile,
             deleteFile: deleteFormFile,
           }}
-          data={{ user: user?.user }}
           formUrl="https://form.adsp-uat.alberta.ca"
         >
           <JsonFormsWrapper definition={definition} data={data} onChange={onChange} readonly={submitting} />
